Remove frontpagereddit fragment due to behavior is the exact same as subreddit fragmetn (but different api calls)
